requirements.txt
asgiref==3.3.1
astroid==2.4.2
colorama==0.4.4
Django==3.1.6
django-crispy-forms==1.11.0
djangorestframework==3.12.2
gunicorn==20.0.4
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
wrapt==1.12.1
whitenoise==5.2.0